Preserve alternate file when loading template
Apparently, :read changes the alternate filename to the file it read
from, causing :b# and similar commands to switch to the template file
right after loading a template.

Fortunately, a :keepalt command already exists to cover this case. It's
put to use when loading the template to avoid exposing internals like
this.