Allow count for :LoadTemplate
A count can now be passed to :LoadTemplate, which will be used as the
range for :read. Default is 0 to retain current behavior by default.

This means :LoadTemplate <template> will work the same way as
before, but :$LoadTemplate <template> is now supported and will load the
template after the last line of the buffer.